requirements.txt
openai-whisper
yt-dlp
sentence-transformers
scikit-learn
numpy
stop-words
